Guessing Game - Visual Basic
This is a guessing game made using Visual Basic and the Windows Form Template. I decided to do this as my first project since I had already made a guessing game in FreeBasic but I wanted to re do it and improve now being able to add an UI and more features.